+++ CHANGES.md
@@ -3,6 +3,7 @@
 #### next release (8.7.12)
 
 - support URL parameters in a GetLegendGraphic request for a layer without a style configured
+- Enhanced error processing for obtaining user location
 - [The next improvement]
 
 #### 8.7.11 - 2024-12-18